fix(storage): extract writer batch/collect helpers to writer/batch.rs (AW-POOL-V9-FIX-3)
writer/mod.rs exceeded the 1000-line decomposition cap after
SerialWriterQueue was folded in. Move writer_loop, process_batch,
process_message_admission_group, process_queued_write, collect_batch,
receive_next_work, process_diagnostic_batch, and copy_error into a new
writer/batch.rs submodule, widening only the items their signatures
require to pub(crate). No behavior change: 136/136 atm-storage-rusqlite
lib tests pass unchanged.
